fix(core): rename host-prefixed files to generic names (rule 31 violations)
codex-cli-fallback.ts -> cli-fallback.ts
codex-thread-key.ts -> thread-key.ts

These generic core modules carried host-specific (Codex) filename prefixes,
violating CLAUDE.md rule 31: 'Core package files must never have host-specific
prefixes — generic modules in core should use generic names. Host adapters
wrap core, not the other way around.'

Symbols keep their CodexCli*/CODEX_THREAD_KEY names (they describe behavior);
only the FILE names (which rule 31 governs) lose the host prefix. All 5
internal import sites and the @remnic/core exports map updated. No cross-
package subpath consumers existed (verified by grep).

Preceding fix for the rule-31 graduation in the next commit (check #15 made
blocking). Per #1528 process: fix violations before the check is hardened so
every commit stays green.

Part of #1528.
